refactored takeaway class adding Order class and moving methods over
